Mild quantities of balancing requests

+++ Content/BehaviorOverrides/BossAIs/Perforators/FallingIchor.cs
@@ -1,4 +1,5 @@
 using CalamityMod;
+using CalamityMod.NPCs.Perforator;
 using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ID;
@@ -17,17 +18,22 @@ namespace InfernumMode.Content.BehaviorOverrides.BossAIs.Perforators
             Projectile.hostile = true;
             Projectile.tileCollide = false;
             Projectile.timeLeft = 420;
-            Projectile.penetrate = 1;
+            Projectile.penetrate = -1;
         }
 
         public override void AI()
         {
-            Projectile.tileCollide = Projectile.timeLeft < 350;
+            Projectile.tileCollide = Projectile.timeLeft < 300;
+
+            bool smallWormIsPresent = NPC.AnyNPCs(ModContent.NPCType<PerforatorHeadSmall>());
+            if (smallWormIsPresent)
+                Projectile.tileCollide = false;
+
             Projectile.rotation = Projectile.velocity.ToRotation() + MathHelper.PiOver2;
 
             bool shouldDie = Collision.SolidCollision(Projectile.position, Projectile.width, Projectile.height);
             shouldDie &= !TileID.Sets.Platforms[CalamityUtils.ParanoidTileRetrieval((int)Projectile.Center.X / 16, (int)Projectile.Center.Y / 16).TileType];
-            if (shouldDie)
+            if (shouldDie && Projectile.tileCollide)
                 Projectile.Kill();
 
             // Release blood idly.
